+++ bitsandbytes/optim/rmsprop.py
@@ -23,11 +23,11 @@ class RMSprop(Optimizer1State):
     ):
         if alpha == 0:
             raise NotImplementedError(
-                f"RMSprop with alpha==0.0 is not supported!"
+                "RMSprop with alpha==0.0 is not supported!"
             )
         if centered:
-            raise NotImplementedError(f"Centered RMSprop is not supported!")
-        super(RMSprop, self).__init__(
+            raise NotImplementedError("Centered RMSprop is not supported!")
+        super().__init__(
             "rmsprop",
             params,
             lr,
@@ -59,11 +59,11 @@ class RMSprop8bit(Optimizer1State):
     ):
         if alpha == 0:
             raise NotImplementedError(
-                f"RMSprop with alpha==0.0 is not supported!"
+                "RMSprop with alpha==0.0 is not supported!"
             )
         if centered:
-            raise NotImplementedError(f"Centered RMSprop is not supported!")
-        super(RMSprop8bit, self).__init__(
+            raise NotImplementedError("Centered RMSprop is not supported!")
+        super().__init__(
             "rmsprop",
             params,
             lr,
@@ -96,11 +96,11 @@ class RMSprop32bit(Optimizer1State):
 
         if alpha == 0:
             raise NotImplementedError(
-                f"RMSprop with alpha==0.0 is not supported!"
+                "RMSprop with alpha==0.0 is not supported!"
             )
         if centered:
-            raise NotImplementedError(f"Centered RMSprop is not supported!")
-        super(RMSprop32bit, self).__init__(
+            raise NotImplementedError("Centered RMSprop is not supported!")
+        super().__init__(
             "rmsprop",
             params,
             lr,